+++ CHANGELOG.md
@@ -2,7 +2,37 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [1.0.1](https://github.com/Snaphy-Automata/paddle-sdk/compare/v1.2.2...v1.0.1) (2019-10-10)
+## [2.2.0](https://github.com/avaly/paddle-sdk/compare/v2.1.0...v2.2.0) (2020-05-12)
+
+
+### Features
+
+* Add TS declaration file and build command ([562de86](https://github.com/avaly/paddle-sdk/commit/562de86cc14399c0bc7152e24cea6176933dd99c))
+
+## [2.1.0](https://github.com/avaly/paddle-sdk/compare/v2.0.0...v2.1.0) (2019-12-29)
+
+
+### Features
+
+* Update subscription plan method ([310f54d](https://github.com/avaly/paddle-sdk/commit/310f54d932b3d4f6715c9ee668cdc6c43523ce4b))
+
+## [2.0.0](https://github.com/avaly/paddle-sdk/compare/v1.3.0...v2.0.0) (2019-12-27)
+
+
+### ⚠ BREAKING CHANGES
+
+* Requires node v10+
+
+### Features
+
+* Upgrade all dependencies ([ce2d240](https://github.com/avaly/paddle-sdk/commit/ce2d2401681d1564e1c77f7904a98de0186fb85a))
+
+## [1.3.0](https://github.com/avaly/paddle-sdk/compare/v1.2.2...v1.3.0) (2019-11-12)
+
+
+### Features
+
+* Add generatePayLink ([#8](https://github.com/avaly/paddle-sdk/issues/8)) ([a09b6c8](https://github.com/avaly/paddle-sdk/commit/a09b6c8))